Start over
We want to cite the first reference where a term was used, which the old
database didn't do. I've started over from scratch to avoid accidentally
mixing references between the two versions.

Additionally, I've added the reference that coined cas1-4, and changed
the reference to the DOI in order to make this a more robust resource.